requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
pydantic~=2.10.6
pandas~=2.2.3
textual~=2.1.2
plotly-express==0.4.1